backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pillow==11.1.0
psutil==6.1.1
PyAudio==0.2.14
pydub==0.25.1
python-engineio==4.11.2
python-socketio==5.12.1
regex==2024.11.6
requests==2.32.3
simple-websocket==1.1.0
SpeechRecognition==3.14.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Wave==0.0.2
Werkzeug==3.1.3
wsproto==1.2.0
